refactor: centralize URL validation in shared module, fix docstring and sentinel.md
Agent-Logs-Url: https://github.com/xNok/infra-bootstrap-tools/sessions/10902ae9-71bc-44df-8b6d-69c227919640
