+++ .jules/sentinel.md
@@ -1,4 +1,4 @@
 ## 2026-05-02 - SSRF and Prompt Injection via Unvalidated Issue URLs
-**Vulnerability:** External input (`github_issue_url`) was directly interpolated into the system prompt for the LLM agent without validation in `agentic/workflows/jules_workflow.py` and `agentic/jules/agent.py`.
+**Vulnerability:** External input (`github_issue_url`) was directly interpolated into the user prompt passed to the LLM agent without validation in `agentic/workflows/jules_workflow.py` and `agentic/jules/agent.py`.
 **Learning:** Passing unsanitized URLs to an LLM exposes the system to prompt injection (where a malicious URL contains instructions overriding the agent's intent) and SSRF (where the LLM agent could be tricked into querying internal or unintended domains).
 **Prevention:** Always validate external URLs using strict parsing (like `urllib.parse`) before embedding them in prompts. Enforce allowed schemes (`https`), specific domains (`github.com`), and strictly validate path structures to ensure only intended inputs reach the LLM.
