+++ agentic/jules/url_validation.py
@@ -1,0 +1,46 @@
+"""
+URL validation utilities for Jules agent.
+
+Centralizes security-sensitive URL validation to prevent prompt injection
+and SSRF attacks when processing external user inputs.
+"""
+
+import urllib.parse
+
+
+def is_valid_github_issue_url(url: str) -> bool:
+    """
+    Validates that the provided URL is a strict GitHub issue URL.
+
+    Accepted format: https://github.com/<owner>/<repo>/issues/<number>
+    Rejects any URL with extra path segments, query parameters, fragments,
+    non-standard ports, or non-https schemes.
+
+    Args:
+        url: The URL string to validate.
+
+    Returns:
+        True if the URL exactly matches the expected GitHub issue format,
+        False otherwise.
+    """
+    try:
+        parsed = urllib.parse.urlparse(url)
+        if parsed.scheme != "https":
+            return False
+        if parsed.hostname != "github.com":
+            return False
+        if parsed.port is not None:
+            return False
+        if parsed.params or parsed.query or parsed.fragment:
+            return False
+        path_parts = parsed.path.strip("/").split("/")
+        if len(path_parts) != 4:
+            return False
+        owner, repo, resource, issue_number = path_parts
+        if not owner or not repo:
+            return False
+        if resource != "issues" or not issue_number.isdigit():
+            return False
+        return True
+    except Exception:
+        return False
